+++ History.md
@@ -6,6 +6,7 @@
 
 *   [#67](https://github.com/pmd/pmd-regression-tester/pull/67): Report contains errors having nil filename
 *   [#68](https://github.com/pmd/pmd-regression-tester/pull/68): Don't generate a dynamic ruleset if not needed
+*   [#69](https://github.com/pmd/pmd-regression-tester/pull/69): Detect single rules with auto-gen-config
 
 ## External Contributions